Working chunk based server

+++ server/src/main/kotlin/dev/blackoutburst/server/network/packets/server/S05SendChunk.kt
@@ -1,13 +1,13 @@
 package dev.blackoutburst.server.network.packets.server
 
-import dev.blackoutburst.game.maths.Vector3f
+import dev.blackoutburst.game.maths.Vector3i
 import dev.blackoutburst.server.network.packets.PacketPlayOut
 import java.nio.ByteOrder
 
 private const val ID: Byte = 0x05
 
 class S05SendChunk(
-    private val position: Vector3f,
+    private val position: Vector3i,
     private val blockData: List<Byte>
 ): PacketPlayOut() {
 
@@ -15,9 +15,9 @@ class S05SendChunk(
         buffer.apply {
             order(ByteOrder.BIG_ENDIAN)
             put(ID)
-            putFloat(position.x)
-            putFloat(position.y)
-            putFloat(position.z)
+            putInt(position.x)
+            putInt(position.y)
+            putInt(position.z)
             blockData.forEach {
                 put(it)
             }
